feat: 유효성 검사와 에러 훅 분리
