+++ src/core/types/standardError.ts
@@ -1,3 +1,8 @@
 export default interface StandardError extends Error {
   status: number;
 }
+
+export interface ErrorData {
+  details?: unknown;
+  message?: string;
+}
