+++ src/lib/hooks/useAuthFormErrors.ts
@@ -1,0 +1,36 @@
+import { useState } from "react";
+import handleAuthInputBlur from "../utils/handleAuthInputBlur";
+
+interface Errors {
+  email: string | null;
+  nickname: string | null;
+  password: string | null;
+  passwordConfirmation: string | null;
+}
+
+interface Form {
+  email: string;
+  nickname: string;
+  password: string;
+  passwordConfirmation: string;
+}
+
+export default function useAuthFormErrors(form?: Form) {
+  const [errors, setErrors] = useState<Errors>({
+    email: null,
+    nickname: null,
+    password: null,
+    passwordConfirmation: null,
+  });
+
+  const handleErrorChange = (key: string, value: string | null) => {
+    setErrors((prev) => ({
+      ...prev,
+      [key]: value,
+    }));
+  };
+
+  const handleBlur = handleAuthInputBlur(handleErrorChange, form);
+
+  return { errors, handleBlur };
+}
