+++ src/lib/utils/handleLoginError.ts
@@ -1,0 +1,6 @@
+import { ErrorData } from "@/core/types/standardError";
+
+export default function handleLoginError(
+  handleErrorChange: (key: string, value: string | null) => void,
+  errorData: ErrorData,
+) {}
